+++ CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes to this project will be documented in this file.
 Le format utilisé se base sur [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 et le versioning du projet respect les règles  [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.6] - 2022-04-20
+### Added
+- Ping database on bootstrap
+### Changed
+- AuthManager lazy loading
+
 ## [0.9.5] - 2022-04-16
 ### Fixed
 - ClassLoader when generated dynamic resource implementation